% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaps.R
\encoding{UTF-8}
\name{adaps}
\alias{adaps}
\alias{adapsBATCH}
\title{adaps and adapsBATCH}
\source{
\enumerate{
   \item r - How can I check if a file is empty? - Stack Overflow answered by Konrad Rudolph and edited by Geekuna Matata on Apr 23 2014. See \url{http://stackoverflow.com/questions/23254002/how-can-i-check-if-a-file-is-empty}.
   \item r - Better error message for stopifnot? - Stack Overflow answered by Andrie on Dec 1 2011. See \url{http://stackoverflow.com/questions/8343509/better-error-message-for-stopifnot}.
   \item inside-R: TclInterface {tcltk}. See \url{http://www.inside-r.org/r-doc/tcltk/.Tcl}.
   \item R TclTk - Open/Save dialogs. See \url{http://www.sciviews.org/_rgui/tcltk/FileOpenSave.html}.
   \item r - read csv files and perform function, then bind together - Stack Overflow answered by bjoseph on Jan 8 2015. See \url{http://stackoverflow.com/questions/27846715/read-csv-files-and-perform-function-then-bind-together}.
   \item multiple output filenames in R - Stack Overflow asked and edited by Gabelins on Feb 1 2013. See \url{http://stackoverflow.com/questions/14651594/multiple-output-filenames-in-r}.
   \item r - Import date-time at a specified timezone, disregard daylights savings time - Stack Overflow answered by James on Nov 4 2011. See \url{http://stackoverflow.com/questions/8004050/import-date-time-at-a-specified-timezone-disregard-daylights-savings-time}.
   \item CST Central Time > What's current time in Central Time Zone now? (USA & Canada). See \url{http://wwp.greenwichmeantime.com/time-zone/usa/central-time/}.
   \item r - Regex return file name, remove path and file extension - Stack Overflow answered and edited by Ananda Mahto on Feb 25 2013. See \url{http://stackoverflow.com/questions/15073753/regex-return-file-name-remove-path-and-file-extension/15073919}.
}
}
\usage{
adaps(file = tk_choose.files(default = "", caption =
  "Select file(s) to open & hold down Ctrl to choose more than 1 file", multi =
  TRUE, filters = matrix(c("ADAPS file", ".rdb", "ADAPS file", ".RDB"), 4, 2,
  byrow = TRUE)), interactive = TRUE, overwrite = overwrite)

adapsBATCH(path = tk_choose.dir(caption =
  "Select directory with the ADAPS .rdb files"), overwrite = overwrite)
}
\arguments{
\item{file}{Input ADAPS .rdb file(s) to be selected through a file dialog.}

\item{interactive}{If interactive is \code{TRUE}, then the user will select the
filenames(s) to use for saving with the file dialog. In order to select
more than one file, the user must hold down the Ctrl (Control) button
while mouse clicking the chosen files. If interactive is \code{FALSE}, then
the user will select the directory, via the directory dialog, to use
for saving and the original filenames will be used.}

\item{overwrite}{If \code{TRUE}, overwrite any existing spreadsheet.}

\item{path}{Directory path of ADAPS .rdb files to be selected through a
directory dialog. The user will be asked where to find the ADAPS .rdb
files & then the user will be asked where to save the ADAPS .xlsx files.}
}
\value{
ADAPS .xlsx file(s) where sheet1 is named after the original
  filename and the exported format is suitable only for viewing in a
  spreadsheet application.
}
\description{
adaps
Opens single or multiple raw ADAPS .rdb file(s) to modify the format and
  then exports the file(s) in .xlsx format. This is done for a single file or
  multiple files that the user selects.
}
\details{
adapsBATCH
Opens raw ADAPS .rdb files, from a directory, to modify the format and then
  exports the files in .xlsx format.  This is done in a BATCH mode
  (whole directory of ADAPS .rdb files).

Both adaps and adapsBATCH functions perform the same processes on the raw
  ADAPS .rdb files: 1) Read in the file and remove the 1st 4 or 5 lines
  depending on whether NTRU data are present or not, 2) create 4 or 5
  columns (depending on NTRU data are present or not) based on the
  1st 4 or 5 lines, and 3) export the modified file in .xlsx format.

The following lines are representative of the .rdb format used in these
  functions.

Line1    DATETIME	ght
Line2	cfs
Line3	fnu
Line4	ntru # may and may not be present
Line5 19D	16N	16N	16N	16N
}
\examples{
\dontrun{
## Examples to change (an) ADAPS .rdb file(s) interactively and non-interactively
library(ie2misc)
adaps() # default where interactive = TRUE
## Follow the file dialog instructions


adaps(interactive = FALSE)
## Follow the file dialog instructions


## Example to change a directory of ADAPS .rdb files
adapsBATCH()
## Follow the file dialog instructions
}

}

