% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ie2misc.R
\docType{package}
\name{ie2misc}
\alias{ie2misc}
\alias{ie2misc-package}
\title{ie2misc: Irucka Embry's miscellaneous USGS functions}
\description{
ie2misc provides many functions that will be useful to people within the
 USGS, while other functions will be useful for others. Functions are
 included for statistical errors (RMSE, NSE, dr, MAPE, MAE, MAD),
 processing .exp and .psf files, a "+" dyadic operator for use with NA,
 a version of linear interpolation for use with NAs, and making it easier
 to view USGS ADAPS and QW files within a spreadsheet application.
}

