% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\encoding{UTF-8}
\name{mape}
\alias{mape}
\title{Mean absolute percent error (MAPE)}
\source{
r - Better error message for stopifnot? - Stack Overflow answered by Andrie on Dec 1 2011. See \url{http://stackoverflow.com/questions/8343509/better-error-message-for-stopifnot}.
}
\usage{
mape(predicted, observed, na.rm = TRUE)
}
\arguments{
\item{predicted}{A numeric vector that contains the predicted data
points (1st parameter)}

\item{observed}{A numeric vector that contains the observed data
points (2nd parameter)}

\item{na.rm}{A logical vector that determines whether the missing
values should be removed or not.}
}
\value{
mean absolute percent error (MAPE) as a numeric vector. The default
        choice is that any NA values will be removed (\code{na.rm = TRUE}). This can be
        changed by specifying \code{na.rm = FALSE}, such as \code{mape(pre, obs, na.rm = FALSE)}.
}
\description{
This function computes the mean absolute percent error (MAPE).
}
\details{
MAPE is expressed as

\deqn{\frac{1}{n} \sum \limits_{i=1}^n{ 100 \frac{\left| X_i - Y_i \right|} {X_i}}}

where \emph{n} is the number of observations,
	where \emph{X} represents the observations, and
	where \emph{Y} represents the predictions.

Below are some points to remember about MAPE from the Ji reference:
\enumerate{
   \item MAPE is "a measure to validate forecast models",
   \item MAPE is "a standardized value and is independent of the unit of
     the measurement",
   \item MAPE is "meaningful only if all \eqn{X_i} values are positive",
   \item MAPE is "unstable when \eqn{X_i} values are near zero", and
   \item "If X and Y are interchanged, the the MAPE will result in a
     different value."
}
}
\examples{
library(ie2misc)
# All of the following examples use the default value of na.rm = TRUE
obs <- 1:10 # observed
pre <- 2:11 # predicted
mape(pre, obs)


require(stats)
set.seed(100) # makes the example reproducible
obs1 <- rnorm(100) # observed
pre1 <- rnorm(100) # predicted


# using the vectors pre1 and obs1
mape(pre1, obs1)


# using a matrix of the numeric vectors pre1 and obs1
mat1 <- matrix(data = c(obs1, pre1), nrow = length(pre1), ncol = 2,
        byrow = FALSE, dimnames = list(c(rep("", length(pre1))),
        c("Predicted", "Observed")))
mape(mat1[, 2], mat1[, 1])

# mat1[, 1] # observed values from column 1 of mat1
# mat1[, 2] # predicted values from column 2 of mat1


# using a data.frame of the numeric vectors pre1 and obs1
df1 <- data.frame(obs1, pre1)
mape(df1[, 2], df1[, 1])

# df1[, 1] # observed values from column 1 of df1
# df1[, 2] # predicted values from column 2 of df1


# using a data.table of the numeric vectors pre1 and obs1
df2 <- data.table(obs1, pre1)
mape(df2[, 2, with = FALSE][[1]], df2[, 1, with = FALSE][[1]])

# df2[, 1, with = FALSE][[1]] # observed values from column 1 of df2
# df2[, 2, with = FALSE][[1]] # predicted values from column 2 of df2



}
\references{
Lei Ji and Kevin Gallo, "An Agreement Coefficient for Image Comparison", \emph{Photogrammetric Engineering & Remote Sensing}, Vol. 72, No. 7, July 2006, p. 823-8335, \url{http://www.asprs.org/a/publications/pers/2006journal/july/2006_jul_823-833.pdf}.
}
\seealso{
\code{\link{mae}} for mean absolute error (MAE), \code{\link{madstat}} for
 mean absolute deviation (MAD), \code{\link{dr}} for index of agreement (dr), \code{\link{vnse}}
 for Nash-Sutcliffe model efficiency (NSE), and \code{\link{rmse}} for
 root mean square error (RMSE).

Other statistical.error.functions: \code{\link{madstat}},
  \code{\link{mae}}, \code{\link{rmse}}, \code{\link{vnse}}
}

