% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qw.R
\encoding{UTF-8}
\name{qw}
\alias{qw}
\alias{qwBATCH}
\title{qw and qwBATCH}
\source{
\enumerate{
   \item r - How can I check if a file is empty? - Stack Overflow answered by Konrad Rudolph and edited by Geekuna Matata on Apr 23 2014. See \url{http://stackoverflow.com/questions/23254002/how-can-i-check-if-a-file-is-empty}.
   \item r - Better error message for stopifnot? - Stack Overflow answered by Andrie on Dec 1 2011. See \url{http://stackoverflow.com/questions/8343509/better-error-message-for-stopifnot}.
   \item inside-R: TclInterface {tcltk}. See \url{http://www.inside-r.org/r-doc/tcltk/.Tcl}.
   \item James Wettenhall & Philippe Grosjean, File Open/Save dialogs in R tcltk, December 01, 2015. See \url{http://www.sciviews.org/recipes/tcltk/TclTk-file-open-save-dialogs/}.
   \item r - read csv files and perform function, then bind together - Stack Overflow answered by bjoseph on Jan 8 2015. See \url{http://stackoverflow.com/questions/27846715/read-csv-files-and-perform-function-then-bind-together}.
   \item r - Convert column classes in data.table - Stack Overflow answered by Matt Dowle on Dec 27 2013. See \url{http://stackoverflow.com/questions/7813578/convert-column-classes-in-data-table}.
   \item Does column exist and how to rearrange columns in R data frame - Stack Overflow answered and edited by Peter McMahan on Aug 2 2009. See \url{http://stackoverflow.com/questions/1177919/does-column-exist-and-how-to-rearrange-columns-in-r-data-frame}.
   \item time - Dealing with timestamps in R - Stack Overflow answered by Dirk Eddelbuettel on Dec 26 2009. See \url{http://stackoverflow.com/questions/1962278/dealing-with-timestamps-in-r/1962336}.
   \item R help - How to change the default Date format for write.csv function? answered by William Dunlap on Dec 28, 2009. See \url{http://r.789695.n4.nabble.com/How-to-change-the-default-Date-format-for-write-csv-function-td989826.html}.
   \item inside-R: strptime {base}. See \url{http://www.inside-r.org/r-doc/base/strftime}.
   \item GMT: Greenwich Mean Time - World Time / Time in every Time Zone: Eastern Time EST (USA & Canada) &gt; What's the current time now in EST? (EDT in Summer). See \url{http://wwp.greenwichmeantime.com/time-zone/usa/eastern-time/}.
   \item GMT: Greenwich Mean Time - World Time / Time in every Time Zone: CST Central Time &gt; What's current time in Central Time Zone now? (USA & Canada). See \url{http://wwp.greenwichmeantime.com/time-zone/usa/central-time/}.
   \item GMT: Greenwich Mean Time - World Time / Time in every Time Zone: Mountain Time (MST - USA & Canada). What's the current time now? / MDT in Summer. See \url{http://wwp.greenwichmeantime.com/time-zone/usa/mountain-time/}.
   \item GMT: Greenwich Mean Time - World Time / Time in every Time Zone: PST: Pacific Time (USA & Canada). What's the current time now in PST? / PDT in Summer. See \url{http://wwp.greenwichmeantime.com/time-zone/usa/pacific-time/}.
   \item multiple output filenames in R - Stack Overflow asked and edited by Gabelins on Feb 1 2013. See \url{http://stackoverflow.com/questions/14651594/multiple-output-filenames-in-r}.
   \item r - Regex return file name, remove path and file extension - Stack Overflow answered and edited by Ananda Mahto on Feb 25 2013. See \url{http://stackoverflow.com/questions/15073753/regex-return-file-name-remove-path-and-file-extension/15073919}.
   \item warnings - How do I get rid of the NOTE's generated by R CMD check when using for example ddply in my package? - Stack Overflow answered by shadow on Mar 4 2015 and edited by shadow on Mar 5 2015. See \url{https://stackoverflow.com/questions/28851812/how-do-i-get-rid-of-the-notes-generated-by-r-cmd-check-when-using-for-example-d}.
}
}
\usage{
qw(file = tk_choose.files(default = "", caption =
  "Select file(s) to open & hold down Ctrl to choose more than 1 file", multi =
  TRUE, filters = matrix(c("QW file", "*"), 1, 2, byrow = TRUE)),
  interactive = TRUE, overwrite = TRUE)

qwBATCH(path = tk_choose.dir(caption = "Select directory with the QW files"),
  pattern = "qwdata$", overwrite = TRUE)
}
\arguments{
\item{file}{Input QW file(s) to be selected through a file dialog.}

\item{interactive}{If interactive is \code{TRUE}, then the user will select the
filenames(s) to use for saving with the file dialog. In order to select
more than one file, the user must hold down the Ctrl (Control) button
while mouse clicking the chosen files. If interactive is \code{FALSE}, then
the user will select the directory, via the directory dialog, to use for
saving and the original filenames will be used.}

\item{overwrite}{If \code{TRUE}, overwrite any existing spreadsheet.}

\item{path}{Directory path of QW files to be selected through a directory
dialog. The user will be asked where to find the QW files & then the
user will be asked where to save the QW .xlsx files.}

\item{pattern}{The default pattern is \code{qwdata$} (the filename ends with a case
insensitive "qwdata"). This pattern can be changed by
  \code{qwBATCH(pattern = "pattern")}.}
}
\value{
QW .xlsx file(s) where sheet1 is named after the original filename.
}
\description{
qw and qwBATCH process raw QW files.
}
\details{
qw function opens a single or multiple raw, QW file(s) to modify the format
  and then exports the file(s) in .xlsx format. This is done for a single
  file or multiple files that the user selects.

qwBATCH function opens raw QW files, from a directory, to modify the format
  and then exports the files in .xlsx format.  This is done in a BATCH mode
  (whole directory of QW files).

Both qw and qwBATCH functions perform the same processes on the raw QW
  files: 1) remove row 1 (which starts with 5s), 2) create a column with
  the Date and Time rather than 2 separate Date and Time columns, and
  3) exports the QW file as a spreadsheet with three sheets (sheet1
  corrects the timezones for Daylight Savings Times and removes the
  timezone datum column, sheet2 does not correct for Daylight Savings Time,
  and sheet3 provides the meta data from the top of the file). The
  supported US timezones are the following: Eastern, Central, Mountain, and
  Pacific. If you are processing stations in the Alaska, Hawai'i and/or
  Puerto Rico timezones, then the timezones will not be adjusted for
  Daylight Savings Time. The same is also true of any stations located in
  other US regions that do not following Daylight Savings Time.
}
\examples{
\dontrun{
# Examples to change (an) QW file(s) interactively and non-interactively
library(ie2misc)
qw() # default where interactive = TRUE
# Follow the file dialog instructions


qw(interactive = FALSE)
# Follow the file dialog instructions


# Example to change QW files in batch mode
qwBATCH()
# Follow the file dialog instructions
}

}

