% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnse.R
\encoding{UTF-8}
\name{vnse}
\alias{vnse}
\title{Nash-Sutcliffe model efficiency (NSE)}
\source{
r - Better error message for stopifnot? - Stack Overflow answered by Andrie on Dec 1 2011. See \url{http://stackoverflow.com/questions/8343509/better-error-message-for-stopifnot}.
}
\usage{
vnse(predicted, observed, na.rm = FALSE)
}
\arguments{
\item{predicted}{A numeric vector that contains the model predicted
data points (1st parameter)}

\item{observed}{A numeric vector that contains the observed data
points (2nd parameter)}

\item{na.rm}{A logical vector that determines whether the missing
values should be removed or not.}
}
\value{
Nash-Sutcliffe model efficiency (NSE) as a numeric vector. The
  default choice is that any NA values will be kept (\code{na.rm = FALSE}). This can
  be changed by specifying \code{na.rm = TRUE}, such as \code{vnse(pre, obs, na.rm = TRUE)}.
}
\description{
This function computes the Nash-Sutcliffe model efficiency (NSE) or
	"Nash and Sutcliffe’s coefficient of efficiency (E)".
}
\details{
NSE or E is expressed as

\deqn{E = 1 - \frac{\sum \limits_{i=1}^n{\left(P_i - O_i\right)^2}}{\sum \limits_{i=1}^n{\left(O_i - \bar{O}\right)^2}}}

where \emph{n} is the number of observations,
where \emph{P} represents the "model estimates or predictions", and
where \emph{O} represents the "pairwise-matched observations that are
  judged to be reliable".

Note: Both P and O should have the same units.

"Nash and Sutcliffe’s coefficient of efficiency (E)" and other
  "dimensionless measures of average error" are fully discussed in the
  Willmott reference.
}
\examples{
library(ie2misc)
obs <- 1:10 # observed
pre <- 2:11 # predicted
vnse(pre, obs)


require(stats)
set.seed(100) # makes the example reproducible
obs1 <- rnorm(100) # observed
pre1 <- rnorm(100) # predicted


# using the vectors pre1 and obs1
vnse(pre1, obs1)


# using a matrix of the numeric vectors pre1 and obs1
mat1 <- matrix(data = c(obs1, pre1), nrow = length(pre1), ncol = 2,
   byrow = FALSE, dimnames = list(c(rep("", length(pre1))),
   c("Predicted", "Observed")))
vnse(mat1[, 2], mat1[, 1])

# mat1[, 1] # observed values from column 1 of mat1
# mat1[, 2] # predicted values from column 2 of mat1


# using a data.frame of the numeric vectors pre1 and obs1
df1 <- data.frame(obs1, pre1)
vnse(df1[, 2], df1[, 1])

# df1[, 1] # observed values from column 1 of df1
# df1[, 2] # predicted values from column 2 of df1


# using a data.table of the numeric vectors pre1 and obs1
df2 <- data.table(obs1, pre1)
vnse(df2[, 2, with = FALSE][[1]], df2[, 1, with = FALSE][[1]])

# df2[, 1, with = FALSE][[1]] # observed values from column 1 of df2
# df2[, 2, with = FALSE][[1]] # predicted values from column 2 of df2



}
\references{
Cort J. Willmott, Scott M. Robeson, and Kenji Matsuura, "A refined index of model performance", \emph{International Journal of Climatology}, Volume 32, Issue 13, pages 2088-2094, 15 November 2012, \url{http://onlinelibrary.wiley.com/doi/10.1002/joc.2419/pdf}.
}
\seealso{
\code{\link{mape}} for mean absolute percent error (MAPE), \code{\link{mae}} for
 mean absolute error (MAE), \code{\link{madstat}} for mean absolute deviation (MAD), \code{\link{dr}}
 for index of agreement (dr), and \code{\link{rmse}} for root mean square error (RMSE).

Other statistical.error.functions: \code{\link{dr}},
  \code{\link{madstat}}, \code{\link{mae}},
  \code{\link{mape}}, \code{\link{rmse}}
}

